src/requirements.txt
bson
pymongo
pandas
scikit-image
google-apputils
python-gflags
glog
tqdm
ipywidgets
opencv-python
gpustat
kaggle-cli
